stub($object)->$method only works if $object->can($method)
Previously, the objects returned from stub($object) were always able
to perform whichever method was called on them. This would be bad:
consider this hypothetical implementation:

    package Production::Code;

    sub fetch_via_lwp {
        my ($class, $user_agent, $url, @parameters) = @_;

        my $response =
            $user_agent->gte($url, @parameters);  # whoops, misspelled ->get as ->gte!

        if ($response->is_success) {
            # ... handle successful response
        } else {
            # ... handle failed response
        }

    }

Imagine a unit test for fetch_via_lwp() which passes in a stubbed-out
$user_agent object:

    my $stub_lwp = stub(LWP::UserAgent->new);

Now if $stub_lwp is passed in to Production::Code->fetch_via_lwp, the bug
in ->fetch_via_lwp will not be exposed during testing: the code will have to
fail in production instead.

This delta modifies Test::Driver::AUTOLOAD to ensure that it throws the same
type of error that perl throws when you call some random non-existant method
on an object.